Take into consideration the PointManager
In all tests and objects

+++ core/headers/PointManager.h
@@ -55,7 +55,7 @@ private:
 	PointManager();
 	~PointManager();
 
-	typedef int IndexType;
+	typedef unsigned long IndexType;
 
 	IndexType m_fromCoordToIndex(float x, float y, float z);
